Added images, README, edited text, made site responsive
Edited the html to make the site a parrot site. Added media queries to make site more responsive
Also added a README file for the project, haven't written any substantial info in the readme though.

This info was meant to be in the last commit but I accidentally messed it up.